+++ R/modify_helpers.R
@@ -1,0 +1,64 @@
+#' Add or modify value
+#'
+#' Modify a value in a row of a data frame based on its value in another column.
+#' If the value is not found, a new row is added.
+#'
+#' @param df Data frame
+#' @param lookup_column Column to look up
+#' @param lookup_column_values Values to look up
+#' @param modify_column Column to modify
+#' @param new_value New value to modify/add to modify_column
+#'
+#' @return A data frame
+modify_or_add_row <- function(df,
+                              lookup_column,
+                              lookup_column_values,
+                              modify_column,
+                              new_value) {
+  if (any(df[[lookup_column]] %in% lookup_column_values)) {
+    df[df[[lookup_column]] %in% lookup_column_values, modify_column] <- new_value
+  } else {
+    df <- dplyr::bind_rows(df,
+                           dplyr::tibble(!!lookup_column := lookup_column_values,
+                                         !!modify_column := new_value
+                           )
+    )
+  }
+
+  return(df)
+}
+
+
+modify_or_add_in_column <- function(df,
+                                    lookup_column,
+                                    lookup_column_values,
+                                    new_value
+                                    ) {
+  if (any(df[[lookup_column]] %in% lookup_column_values)) {
+    df[df[[lookup_column]] %in% lookup_column_values, lookup_column] <- new_value
+  } else {
+    df <- dplyr::bind_rows(df,
+                           dplyr::tibble(!!lookup_column := new_value)
+                           )
+  }
+
+  return(df)
+}
+
+modify_table1 <- function(x, keyword, value) {
+  x$table1 <- modify_or_add_row(x$table1, "keyword", keyword, "value", value)
+  return(x)
+}
+
+remove_keyword_table1 <- function(x, keyword) {
+  x$table1 <- x$table1 %>%
+    dplyr::filter(keyword != !!keyword)
+
+  return(x)
+}
+
+get_table1_value <- function(x, keyword) {
+  x$table1 %>%
+    dplyr::filter(keyword == !!keyword) %>%
+    dplyr::pull(value)
+}